Dockerfile: Specify fully qualified image name
Why:
podman and containerd require fully qualified names for Docker images,
meaning that the hostname of the registry needs to be present. Docker
defaults to docker.io when that is missing, but both podman and
containerd require specific configuration for that to happen on the part
of the operator.

What:
Make life easier for users of alternate container engines by specifying
the fully qualified names for images in the Dockerfile

Signed-off-by: Alexandros Kosiaris <akosiaris@gmail.com>