ships.clj: Ignore a :not-empty? warning by clj-kondo
Why:
Given clj-kondo doesn't complain about the rest of the Clojure codebase
after the last few patches, this is the last hanging fruit. I 've been
pondering over this for a while. Replacing (not (empty?) with idiom
(seq) as per https://clojuredocs.org/clojure.core/empty_q feels a bit
weird. It seems I am not the only one with this opinion per

  https://github.com/clj-kondo/clj-kondo/issues/937

Given the above, I am inclined to just ignore it for now, but happy to
be convinced otherwise

What:
Just instruct clj-kondo to ignore this